pull flag check up to top level
