+++ app/jenkins-for-jira-ui/src/components/JenkinsServerList/JenkinsServerList.tsx
@@ -17,10 +17,7 @@ import {
 	AnalyticsUiEventsEnum
 } from '../../common/analytics/analytics-events';
 import { redirectFromGetStarted } from '../../api/redirectFromGetStarted';
-import { FeatureFlags } from '../../hooks/useFeatureFlag';
 import { fetchCloudId } from '../../api/fetchCloudId';
-import { fetchFeatureFlagFromBackend } from '../../api/fetchFeatureFlagFromBackend';
-import { TopPanel } from './TopPanel/TopPanel';
 
 const JenkinsServerList = (): JSX.Element => {
 	const history = useHistory();
@@ -28,7 +25,6 @@ const JenkinsServerList = (): JSX.Element => {
 	const [jenkinsServers, setJenkinsServers] = useState<JenkinsServer[]>();
 	const [moduleKey, setModuleKey] = useState<string>();
 	const [cloudId, setCloudId] = useState<string>();
-	const [renovateConfigFlag, setRenovateConfigFlag] = useState<boolean>(false);
 
 	const fetchAllJenkinsServers = async () => {
 		const servers = await getAllJenkinsServers() || [];
@@ -45,19 +41,11 @@ const JenkinsServerList = (): JSX.Element => {
 		setModuleKey(currentModuleKey);
 	}, []);
 
-	const fetchFeatureFlag = useCallback(async () => {
-		const renovatedJenkinsFeatureFlag = await fetchFeatureFlagFromBackend(
-			FeatureFlags.RENOVATED_JENKINS_FOR_JIRA_CONFIG_FLOW
-		);
-		setRenovateConfigFlag(renovatedJenkinsFeatureFlag);
-	}, []);
-
 	useEffect(() => {
 		fetchAllJenkinsServers();
 		redirectToAdminPage();
 		getCloudId();
-		fetchFeatureFlag();
-	}, [redirectToAdminPage, cloudId, fetchFeatureFlag]);
+	}, [redirectToAdminPage, cloudId]);
 
 	if (!jenkinsServers || !moduleKey || moduleKey === 'get-started-page') {
 		return <JenkinsSpinner secondaryClassName={spinnerHeight} />;
@@ -80,14 +68,10 @@ const JenkinsServerList = (): JSX.Element => {
 	const pageHeaderActions = (
 		<ButtonGroup>
 			<Button appearance="primary" onClick={() => onClickConnect()}>
-				{renovateConfigFlag ? 'Connect a new Jenkins server' : 'Connect a Jenkins server'}
+				Connect a Jenkins server
 			</Button>
-			{/* TODO - add onClick event */}
-			{renovateConfigFlag && <Button>Share page</Button>}
 		</ButtonGroup>
 	);
-
-	const headerText = renovateConfigFlag ? 'Jenkins for Jira' : 'Jenkins configuration';
 
 	let contentToRender;
 
@@ -95,17 +79,16 @@ const JenkinsServerList = (): JSX.Element => {
 		contentToRender = (
 			<>
 				<div className={headerContainer}>
-					<PageHeader actions={pageHeaderActions}>{headerText}</PageHeader>
+					<PageHeader actions={pageHeaderActions}>Jenkins configuration</PageHeader>
 				</div>
 
-				{!renovateConfigFlag && <StyledDescription>
+				<StyledDescription>
 					After you connect your Jenkins server to Jira and send a deployment
 					event from your CI/CD tool, you will be able to view development
 					information within your linked Jira issue and view deployment pipelines
 					over a timeline with insights.
-				</StyledDescription>}
+				</StyledDescription>
 
-				{renovateConfigFlag && <TopPanel />}
 				<ConnectedServers jenkinsServerList={jenkinsServers} refreshServers={fetchAllJenkinsServers} />
 			</>
 		);
